+++ csv-extractor/csv_extractor.py
@@ -2,19 +2,20 @@
 """
 A module for extracting datapoints from CSV files.
 """
+import asyncio
 import csv
 import logging
 import sys
-import threading
 import time
 from collections import defaultdict
-from itertools import chain
 from operator import itemgetter
-from typing import Dict
 
 from cognite.client import APIError
 from cognite.client.stable.datapoints import Datapoint, TimeseriesWithDatapoints
 from cognite.client.stable.time_series import TimeSeries
+
+import aiofiles
+import aiofiles.os
 
 logger = logging.getLogger(__name__)
 
@@ -22,7 +23,9 @@ logger = logging.getLogger(__name__)
 BATCH_MAX = 1000  # Maximum number of time series batched at once
 
 
-def extract_data_points(client, monitor, time_series_cache, live_mode: bool, folder_path, failed_path, finished_path):
+async def extract_data_points(
+    client, monitor, time_series_cache, live_mode: bool, folder_path, failed_path, finished_path
+):
     """Find and publish all data points in files found in 'folder_path'.
 
     In `live_mode` will process only 20 newest files, and the search for new files again.
@@ -37,7 +40,7 @@ def extract_data_points(client, monitor, time_series_cache, live_mode: bool, fol
 
         if files:
             files = files[:20] if live_mode else files  # We only process 20 newest before we look again for live
-            process_files(client, monitor, files, time_series_cache, failed_path, finished_path)
+            await process_files(files, client, monitor, time_series_cache, failed_path, finished_path)
 
         if live_mode:
             time.sleep(8)
@@ -91,134 +94,105 @@ def create_data_points(values, timestamps):
     return data_points
 
 
-def create_time_series(client, name: str, external_id: str) -> None:
-    """Create a new time series when the 'external_id' isn't found."""
-    new_time_series = TimeSeries(
-        name=name, description="Auto-generated time series, external ID not found", metadata={"externalID": external_id}
-    )
-    _log_error(client.time_series.post_time_series, [new_time_series])
+async def process_files(paths, client, monitor, time_series_cache, failed_path, finished_path, batch_size=10):
+    """Process a batch of csv file at a time, and either delete or move them when done."""
+    monitor.successfully_processed_files_gauge.set(0)
+    monitor.unprocessed_files_gauge.set(len(paths))
+
+    for i in range(0, len(paths), batch_size):
+        start_time = time.time()
+
+        futures = []
+        for path in paths[i : i + batch_size]:
+            futures.append(process_one_file(path, client, time_series_cache, monitor, failed_path, finished_path))
+        await asyncio.gather(*futures, return_exceptions=True)
+
+        logger.info("Total time to process {} files: {:.2f} seconds".format(len(futures), time.time() - start_time))
 
 
-def get_parsed_file(path) -> Dict[str, list]:
+async def process_one_file(path, client, time_series_cache, monitor, failed_path, finished_path):
+    """Read and understand content of one csv file, then send data points to CDF."""
+    start_time = time.time()
+
+    try:
+        content = await read_and_parse_csv_file(path)
+
+        data_points, new_time_series, data_points_count, time_series_count = convert_file_content_to_data_points(
+            content, time_series_cache
+        )
+        if new_time_series:
+            _log_error(client.time_series.post_time_series, new_time_series)
+            monitor.incr_created_time_series_counter(len(new_time_series))
+
+        _log_error(client.datapoints.post_multi_time_series_datapoints, data_points)
+
+        if finished_path is None:
+            await aiofiles.os.remove(path)
+        else:
+            await aiofiles.os.rename(path, finished_path.joinpath(path.name))
+
+        logger.info("Time to process file {}: {:.2f} seconds".format(path, time.time() - start_time))
+
+    except IOError as exc:
+        logger.debug("Unable to open or delete file {}: {!s}".format(path, exc))
+    except Exception as exc:
+        logger.error("Parsing of file {} failed: {!s}".format(path, exc), exc_info=exc)
+        monitor.incr_failed_files_counter()
+        if failed_path is not None:
+            await aiofiles.os.rename(path, failed_path.joinpath(path.name))
+    else:
+        monitor.successfully_processed_files_gauge.inc()
+        monitor.incr_total_data_points_counter(data_points_count)
+        monitor.count_of_time_series_gauge.set(time_series_count)
+
+    monitor.unprocessed_files_gauge.dec()
+    monitor.push()
+
+
+async def read_and_parse_csv_file(csv_path):
     """Parse the csv file and return the data in a {col_name -> list_of_row_items} dictionary"""
     parsed_file = defaultdict(list)
-    with open(path, "r", encoding="latin-1") as f:
-        data = csv.DictReader(f, delimiter=";")
-        for row in data:
-            for k, v in row.items():
-                parsed_file[k].append(v)
+    async with aiofiles.open(csv_path, mode="r", encoding="latin-1") as f:
+        lines = await f.readlines()
+    for row in csv.DictReader(lines, delimiter=";"):
+        for k, v in row.items():
+            parsed_file[k].append(v)
     return parsed_file
 
 
-def process_csv_file(client, monitor, csv_path, existing_time_series):
-    start_time = time.time()
+def convert_file_content_to_data_points(content, existing_time_series):
+    """Take content of a csv file and produce list of TimeseriesWithDatapoints"""
+    timestamps = content[""][1:]  # ignore garbage value in first line
+    del content[""]  # remove the timestamps from the dictionary
 
-    parsed_file = get_parsed_file(csv_path)
-    timestamps = parsed_file[""][1:]  # ignore garbage value in first line
-    del parsed_file[""]  # remove the timestamps from the dictionary
-
-    count_of_data_points = 0
+    time_series_with_data_points = []
+    unknown_time_series = []
     unique_external_ids = set()  # Count number of time series processed
-    current_time_series = []  # List of time series being processed
-    network_threads = []
-    for col_name, v in parsed_file.items():
-        if len(current_time_series) >= 1000:
-            network_threads.append(
-                threading.Thread(
-                    target=_log_error,
-                    args=(client.datapoints.post_multi_time_series_datapoints, current_time_series[:]),
+    count_of_data_points = 0
+
+    for col_name, values in content.items():
+        name, _, external_id = col_name.rpartition(":")
+        name, external_id = name.strip(), external_id.strip()
+
+        if external_id not in existing_time_series:
+            existing_time_series[external_id] = name
+            unknown_time_series.append(
+                TimeSeries(
+                    name=name,
+                    description="Auto-generated time series, external ID not found",
+                    metadata={"externalID": external_id},
                 )
             )
 
-            current_time_series.clear()
-
-        name = col_name.rpartition(":")[2].strip()
-        external_id = col_name.rpartition(":")[0].strip()
-
-        if external_id not in existing_time_series:
-            create_time_series(client, name, external_id)
-            existing_time_series[external_id] = name
-            monitor.incr_created_time_series_counter()
-
-        data_points = create_data_points(v[1:], timestamps)
+        data_points = create_data_points(values[1:], timestamps)
         if data_points:
-            current_time_series.append(
+            time_series_with_data_points.append(
                 TimeseriesWithDatapoints(name=existing_time_series[external_id], datapoints=data_points)
             )
             count_of_data_points += len(data_points)
             unique_external_ids.add(external_id)
 
-    if current_time_series:
-        network_threads.append(
-            threading.Thread(
-                target=_log_error, args=(client.datapoints.post_multi_time_series_datapoints, current_time_series)
-            )
-        )
-
-    logger.info("Time to process file {}: {:.2f} seconds".format(csv_path, time.time() - start_time))
-
-    return network_threads, count_of_data_points, len(unique_external_ids)
-
-
-def post_all_data(queue, monitor, finished_path):
-    start_time = time.time()
-
-    all_threads = list(chain(*map(itemgetter(0), queue)))
-    for thread in all_threads:
-        thread.start()
-    for thread in all_threads:
-        thread.join()
-
-    for path in map(itemgetter(1), queue):
-        try:
-            if finished_path is None:
-                path.unlink()
-            else:
-                path.replace(finished_path.joinpath(path.name))
-        except IOError as exc:
-            logger.debug("Unable to delete file {}: {!s}".format(path, exc))
-
-    monitor.incr_total_data_points_counter(sum(map(itemgetter(2), queue)))
-
-    logger.info("Total time to send batch of request: {:.2f} seconds".format(time.time() - start_time))
-
-
-def process_files(client, monitor, paths, time_series_cache, failed_path, finished_path) -> None:
-    """Process one csv file at a time, and either delete it or move it when done."""
-    monitor.successfully_processed_files_gauge.set(0)
-    monitor.unprocessed_files_gauge.set(len(paths))
-    thread_queue = []
-    start_time = time.time()
-
-    for path in paths:
-        try:
-            threads, data_points_count, time_series_count = process_csv_file(client, monitor, path, time_series_cache)
-            thread_queue.append((threads, path, data_points_count))
-        except IOError as exc:
-            logger.debug("Unable to open file {}: {!s}".format(path, exc))
-        except Exception as exc:
-            logger.error("Parsing of file {} failed: {!s}".format(path, exc), exc_info=exc)
-            monitor.incr_failed_files_counter()
-
-            if failed_path is not None:
-                path.replace(failed_path.joinpath(path.name))
-
-        else:
-            monitor.successfully_processed_files_gauge.inc()
-            monitor.count_of_time_series_gauge.set(time_series_count)
-
-            if len(thread_queue) >= 20:
-                post_all_data(thread_queue, monitor, finished_path)
-                thread_queue.clear()
-
-        monitor.unprocessed_files_gauge.dec()
-        monitor.push()
-
-    if thread_queue:
-        post_all_data(thread_queue, monitor, finished_path)
-        monitor.push()
-
-    logger.info("Total time to process {} of files: {:.2f} seconds".format(len(paths), time.time() - start_time))
+    return time_series_with_data_points, unknown_time_series, count_of_data_points, len(unique_external_ids)
 
 
 def find_files_in_path(folder_path):