+++ csv-extractor/tests/test_extractor.py
@@ -14,7 +14,7 @@ class TestExtractor:
     folder_path = Path(__file__).parent / "test_files"
 
     def test_find_files_in_path_historical(self):
-        result = find_files_in_path(Path(self.folder_path), 0)
+        result = find_files_in_path(Path(self.folder_path))
         assert len(result), len(os.listdir(self.folder_path))
 
     def test_create_datapoints(self):